+++ hdfs/src/main/java/org/testifyproject/resource/hdfs/MiniDFSResource.java
@@ -19,19 +19,19 @@ import java.io.IOException;
 import org.apache.hadoop.hdfs.DistributedFileSystem;
 import org.apache.hadoop.hdfs.HdfsConfiguration;
 import org.apache.hadoop.hdfs.MiniDFSCluster;
-import org.testifyproject.ResourceInstance;
-import org.testifyproject.ResourceProvider;
+import org.testifyproject.LocalResourceProvider;
+import org.testifyproject.LocalResourceInstance;
 import org.testifyproject.TestContext;
-import org.testifyproject.core.ResourceInstanceBuilder;
+import org.testifyproject.core.LocalResourceInstanceBuilder;
 import org.testifyproject.core.util.FileSystemUtil;
 
 /**
- * An implementation of ResourceProvider that provides a local HDFS test cluster
+ * An implementation of LocalResourceProvider that provides a local HDFS test cluster
  * server and file system client.
  *
  * @author saden
  */
-public class MiniDFSResource implements ResourceProvider<HdfsConfiguration, MiniDFSCluster, DistributedFileSystem> {
+public class MiniDFSResource implements LocalResourceProvider<HdfsConfiguration, MiniDFSCluster, DistributedFileSystem> {
 
     private final FileSystemUtil fileSystemUtil = FileSystemUtil.INSTANCE;
     private MiniDFSCluster hdfsCluster;
@@ -48,7 +48,7 @@ public class MiniDFSResource implements ResourceProvider<HdfsConfiguration, Mini
     }
 
     @Override
-    public ResourceInstance<MiniDFSCluster, DistributedFileSystem> start(TestContext testContext, HdfsConfiguration config) {
+    public LocalResourceInstance<MiniDFSCluster, DistributedFileSystem> start(TestContext testContext, HdfsConfiguration config) {
         try {
             String hdfsDirectory = config.get(MiniDFSCluster.HDFS_MINIDFS_BASEDIR);
             fileSystemUtil.recreateDirectory(hdfsDirectory);
@@ -57,8 +57,8 @@ public class MiniDFSResource implements ResourceProvider<HdfsConfiguration, Mini
             hdfsCluster = builder.build();
             fileSystem = hdfsCluster.getFileSystem();
 
-            return new ResourceInstanceBuilder<MiniDFSCluster, DistributedFileSystem>()
-                    .server(hdfsCluster, "hdfsMiniCluster")
+            return LocalResourceInstanceBuilder.builder()
+                    .resource(hdfsCluster, "hdfsMiniCluster")
                     .client(fileSystem, "hdfsFileSystem")
                     .build();
         } catch (Exception e) {