+++ hdfs/src/test/java/org/testifyproject/resource/hdfs/MiniDFSResourceTest.java
@@ -24,14 +24,14 @@ import org.apache.hadoop.hdfs.DistributedFileSystem;
 import org.apache.hadoop.hdfs.HdfsConfiguration;
 import org.apache.hadoop.hdfs.MiniDFSCluster;
 import static org.assertj.core.api.Assertions.assertThat;
+import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
-import org.testifyproject.ResourceInstance;
+import org.testifyproject.LocalResourceInstance;
 import org.testifyproject.TestContext;
 import org.testifyproject.annotation.Cut;
-import org.testifyproject.annotation.Fixture;
 import org.testifyproject.junit4.UnitTest;
 
 /**
@@ -42,8 +42,12 @@ import org.testifyproject.junit4.UnitTest;
 public class MiniDFSResourceTest {
 
     @Cut
-    @Fixture(destroy = "stop")
     MiniDFSResource cut;
+
+    @After
+    public void destory() {
+        cut.stop();
+    }
 
     @Test
     public void callToStartResourceShouldReturnRequiredResource() throws IOException {
@@ -53,11 +57,11 @@ public class MiniDFSResourceTest {
         HdfsConfiguration config = cut.configure(testContext);
         assertThat(config).isNotNull();
 
-        ResourceInstance<MiniDFSCluster, DistributedFileSystem> result = cut.start(testContext, config);
+        LocalResourceInstance<MiniDFSCluster, DistributedFileSystem> result = cut.start(testContext, config);
 
         assertThat(result).isNotNull();
         assertThat(result.getClient()).isPresent();
-        assertThat(result.getServer()).isNotNull();
+        assertThat(result.getResource()).isNotNull();
 
         FileSystem fileSystem = result.getClient().get().getInstance();
         short replicationFactor = 2;