+++ hsql/src/test/java/org/testifyproject/resource/hsql/InMemoryHSQLResourceTest.java
@@ -19,14 +19,14 @@ import java.sql.Connection;
 import javax.sql.DataSource;
 import static org.assertj.core.api.Assertions.assertThat;
 import org.hsqldb.jdbc.JDBCDataSource;
+import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
-import org.testifyproject.ResourceInstance;
+import org.testifyproject.LocalResourceInstance;
 import org.testifyproject.TestContext;
 import org.testifyproject.annotation.Cut;
-import org.testifyproject.annotation.Fixture;
 import org.testifyproject.junit4.UnitTest;
 
 /**
@@ -37,8 +37,12 @@ import org.testifyproject.junit4.UnitTest;
 public class InMemoryHSQLResourceTest {
 
     @Cut
-    @Fixture(destroy = "stop")
     InMemoryHSQLResource cut;
+
+    @After
+    public void destory() {
+        cut.stop();
+    }
 
     @Test
     public void configureAndStartRequiredResource() {
@@ -48,8 +52,8 @@ public class InMemoryHSQLResourceTest {
         JDBCDataSource config = cut.configure(testContext);
         assertThat(config).isNotNull();
 
-        ResourceInstance<DataSource, Connection> resourceInstance = cut.start(testContext, config);
-        assertThat(resourceInstance.getServer()).isNotNull();
+        LocalResourceInstance<DataSource, Connection> resourceInstance = cut.start(testContext, config);
+        assertThat(resourceInstance.getResource()).isNotNull();
         assertThat(resourceInstance.getClient()).isPresent();
     }