+++ zookeeper/src/test/java/org/testifyproject/resource/zookeeper/ZooKeeperResourceTest.java
@@ -18,14 +18,14 @@ package org.testifyproject.resource.zookeeper;
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.test.TestingServer;
 import static org.assertj.core.api.Assertions.assertThat;
+import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
-import org.testifyproject.ResourceInstance;
+import org.testifyproject.LocalResourceInstance;
 import org.testifyproject.TestContext;
 import org.testifyproject.annotation.Cut;
-import org.testifyproject.annotation.Fixture;
 import org.testifyproject.junit4.UnitTest;
 
 /**
@@ -36,8 +36,12 @@ import org.testifyproject.junit4.UnitTest;
 public class ZooKeeperResourceTest {
 
     @Cut
-    @Fixture(destroy = "stop")
     ZooKeeperResource cut;
+
+    @After
+    public void destory() {
+        cut.stop();
+    }
 
     @Test
     public void callToStartResourceShouldReturnRequiredResource() throws Exception {
@@ -47,11 +51,11 @@ public class ZooKeeperResourceTest {
         Void config = cut.configure(testContext);
         assertThat(config).isNull();
 
-        ResourceInstance<TestingServer, CuratorFramework> result = cut.start(testContext, null);
+        LocalResourceInstance<TestingServer, CuratorFramework> result = cut.start(testContext, null);
 
         assertThat(result).isNotNull();
         assertThat(result.getClient()).isPresent();
-        assertThat(result.getServer()).isNotNull();
+        assertThat(result.getResource()).isNotNull();
 
         CuratorFramework client = result.getClient().get().getInstance();
         String testPath = client.create().forPath("/test", "test".getBytes());